+++ crates/module-system/module-implementations/sov-sequencer-registry/src/call.rs
@@ -309,9 +309,9 @@ impl<S: sov_modules_api::Spec, Da: sov_modules_api::DaSpec> SequencerRegistry<S,
     ///
     /// # Safety note:
     /// This method panics if:
-    /// - The sequencer is not registered (this should be checked in the `begin_batch_hook` which should always be called before this method).
+    /// - The sequencer is not registered.
     /// - The module account does not have enough funds to pay for the reward (the module balance should be populated in the `GasEnforcer` capability hook).
-    pub(crate) fn reward_sequencer(
+    pub fn reward_sequencer(
         &self,
         sequencer: &Da::Address,
         amount: u64,
@@ -320,7 +320,11 @@ impl<S: sov_modules_api::Spec, Da: sov_modules_api::DaSpec> SequencerRegistry<S,
         let AllowedSequencer {
             address: rollup_address,
             balance: _,
-        } = self.allowed_sequencers.get(sequencer, state).unwrap_infallible().expect("Sequencer must be allowed. This should have been checked in the `begin_batch_hook`. This is a bug");
+        } = self
+            .allowed_sequencers
+            .get(sequencer, state)
+            .unwrap_infallible()
+            .expect("Sequencer must be allowed.");
 
         self.bank
             .transfer_from(